Fix clang build

+++ src/n64/vr4300/exceptions.cpp
@@ -172,10 +172,10 @@ template<MemOp mem_op> void TlbMissException(u64 bad_vaddr)
 {
     if constexpr (log_exceptions) log(std::format("EXCEPTION: TLB Miss; vaddr {:016X}", bad_vaddr));
     static constexpr s32 base_addr[2][2] = {
-        0x8000'0000,
-        0x8000'0180,
-        0xBFC0'0200,
-        0xBFC0'0380,
+        s32(0x8000'0000),
+        s32(0x8000'0180),
+        s32(0xBFC0'0200),
+        s32(0xBFC0'0380),
     };
     u64 vector = base_addr[cop0.status.bev][cop0.status.exl];
     HandleException(); // changes status.exl
@@ -219,10 +219,10 @@ template<MemOp mem_op> void XtlbMissException(u64 bad_vaddr)
 {
     if constexpr (log_exceptions) log(std::format("EXCEPTION: XTLB Miss; vaddr {:016X}", bad_vaddr));
     static constexpr s32 base_addr[2][2] = {
-        0x8000'0080,
-        0x8000'0180,
-        0xBFC0'0280,
-        0xBFC0'0380,
+        s32(0x8000'0080),
+        s32(0x8000'0180),
+        s32(0xBFC0'0280),
+        s32(0xBFC0'0380),
     };
     u64 vector = base_addr[cop0.status.bev][cop0.status.exl];
     HandleException(); // changes status.exl
